type_check try to avoid flooding the user with errors originating from a previous type check failure.
When a type check is failed, in order to try and recover to provide further diagnostics when the result type is
dependent on the input types, the stack is "padded" out with Unknown to the depth it would have been if the
operation had succeded.

Previously, subsequent type checks would fail because they didn't support Unknown, but would emit an error
even if they would have succeeded had the original error not occurred.