OpCode: Make procedure calls non-const.
The issue here is that a const may depend on a proc, and that proc may depend on that const.
This potential circular dependency makes things a pain in the ass, because we can't process procs
until we've evaluated consts, but also can't process consts because they may depend on a proc.

A potential solution to this would be lazy evaluation of consts, but that makes cycle-detection
even more awkward.