OpCode: Make UnresolvedIdent's fields more obvious.
Previously, the fields had a required primary token, and an optional secondary token, where the primary token was
either a proc name if there was no secondary token, and a module name if there was a secondary token.

Now it has a required proc token, and an optional module token.